Refactor workshop final project
